// source/frontend/src/models/api/generated/ApiGen_Concepts_DocumentMetadataUpdate.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */

// LINK: @backend/apimodels/Models/Concepts/Document/DocumentMetadataUpdateModel.cs
export interface ApiGen_Concepts_DocumentMetadataUpdate {
  id: number;
  metadataTypeId: number;
  value: string | null;
}