// source/frontend/src/models/api/generated/ApiGen_Mayan_DocumentTypeMetadataType.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { ApiGen_Mayan_DocumentType } from './ApiGen_Mayan_DocumentType';
import { ApiGen_Mayan_MetadataType } from './ApiGen_Mayan_MetadataType';

// LINK: @backend/apimodels/Models/Mayan/Document/DocumentTypeMetadataTypeModel.cs
export interface ApiGen_Mayan_DocumentTypeMetadataType {
  id: number;
  document_type: ApiGen_Mayan_DocumentType | null;
  metadata_type: ApiGen_Mayan_MetadataType | null;
  required: boolean;
  url: string | null;
}